Vagrantfile: Clear locale specific environment variables.
Having LANG and LC_CTYPE set implies that corresponding locale specific
support is available in the VM which may not always be the case
depending on the locale type in question. Also we don't necessarily want
to get caught up with locale specific quagmires anyway.

Clearing these silences the ugly complaints from perl (which git
submodule relies on) in LISA shell.

Tested with OS X.