lisa_shell: Use eth0 as the default NETIF with vagrant.
Using lo as the default NETIF with lisa-ipython silently fails to
provide a working port forwarding arrangement between the host and the
guest.

This crude fix uses eth0 as the default NETIF if $USER is 'vagrant'
which really should be OK (both the username check as well as the
assumption that eth0 is the interface name).